# llamaville/models.py
# generated by fastapi-codegen:
#   filename:  openapi-2.json
#   timestamp: 2021-05-26T13:18:58+00:00

from __future__ import annotations

from datetime import datetime
from typing import Any, Dict, List, Optional

from pydantic import BaseModel, Field


class Warta(BaseModel):
    lamaID: Optional[int] = None
    poczatekWarty: Optional[datetime] = Field(
        None,
        description="Date format according to ISO_8601 for example: yyyy-MM-dd'T'HH:mm:ss.SSSZ",
    )
    koniecWarty: Optional[datetime] = Field(
        None,
        description="Date format according to ISO_8601 for example: yyyy-MM-dd'T'HH:mm:ss.SSSZ",
    )


class Harmonogram(BaseModel):
    Warta: Optional[Warta] = None


class Lama(BaseModel):
    id: Optional[int] = None
    imie: Optional[str] = None
    wiek: Optional[int] = None
    rasa: Optional[Dict[str, Any]] = None
    kolor: Optional[Dict[str, Any]] = None
    kubraczek: Optional[str] = None


class CreateRequest(BaseModel):
    lama: Lama


class UpdateRequest(BaseModel):
    lama: Lama


class LlamaListResponse(BaseModel):
    llamaList: List[Lama]


class AddScheduleRequest(BaseModel):
    schedule: Harmonogram


class ScheduleListResponse(BaseModel):
    llamaList: Optional[List[Harmonogram]] = None


class Exception(BaseModel):
    code: Optional[int] = None
    detail: Optional[str] = None