# openapi.yaml
openapi: 3.0.1
info:
  title: LlamaVille - system zarządzania lamami
  description: System uporządzania lam na posterunku
  termsOfService: https://brzydakville.space
  contact:
    email: llamanager@brzydakville.space
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
- url: https://brzydakville.space/v2
tags:
- name: lama
  description: Panel operacji CRUD na lamach
  externalDocs:
    description: Lamy są fajne, tylko trzeba o nie dbać
    url: https://brzydakville.space
paths:
  /lama:
    get:
      tags:
      - lama
      summary: Pobierz listę lam
      description: Pobierz listę wszystkich lam w bazie
      operationId: getAllLlamas
      responses:
        201:
          description: Return list of llamas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlamaListResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
      security:
      - bville_auth:
        - write_lama
        - read_lama
    post:
      tags:
      - lama
      summary: Dodaj lamię
      description: Wymagane uwierzytelnienie strażnika lam
      operationId: registerLama
      parameters:
      - name: X-HMAC-SIGNATURE
        in: header
        description: integrity verification
        required: true
        schema:
          type: string
      requestBody:
        description: Created lama object
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/CreateRequest'
        required: true
      responses:
        204:
          description: Lama successfully added
          content: {}
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        422:
          description: 'Unprocessable entity. Codes: LLAMA_ALREADY_EXISTS'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
      security:
      - bville_auth:
        - write_lama
        - read_lama
      x-codegen-request-body-name: body
  /lama/{id}:
    put:
      tags:
      - lama
      summary: Zaktualizuj dane lamy
      description: Zaktualizuj dane lamy
      operationId: updateLlama
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-HMAC-SIGNATURE
        in: header
        description: integrity verification
        required: true
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/UpdateRequest'
        required: true
      responses:
        204:
          description: Llama successfully updated
          content: {}
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        422:
          description: 'Unprocessable entity. Codes: LLAMA_DOES_NOT_EXIST'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
      security:
      - bville_auth:
        - write_lama
        - read_lama
      x-codegen-request-body-name: body
    delete:
      tags:
      - lama
      summary: Usuń lamę z bazy danych
      description: Usunięcie lamy z bazy danych
      operationId: deleteLlama
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        204:
          description: Llama successfully deleted
          content: {}
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        422:
          description: 'Unprocessable entity. Codes: LLAMA_DOES_NOT_EXIST'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
      security:
      - bville_auth:
        - write_lama
        - read_lama
  /posterunek:
    get:
      tags:
      - harmonogram
      summary: Pobierz listę wszystkich pozycji w harmonogramie z przypisanymi lamami
      description: Pobierz listę wszystkich pozycji w harmonogramie z przypisanymi
        lamami
      operationId: getSchedule
      responses:
        201:
          description: Return list of all schedule entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleListResponse'
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
      security:
      - bville_auth:
        - write_lama
        - read_lama
    post:
      tags:
      - harmonogram
      summary: Przypisz lamę do posterunku w określonym harmonogramie
      description: Wymagane uwierzytelnienie strażnika lam
      operationId: registerSchedule
      parameters:
      - name: X-HMAC-SIGNATURE
        in: header
        description: integrity verification
        required: true
        schema:
          type: string
      requestBody:
        description: Created schedule object
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/AddScheduleRequest'
        required: true
      responses:
        204:
          description: Schedule successfully added
          content: {}
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
        422:
          description: 'Unprocessable entity. Codes: Schedule time already taken by
            different llama!'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exception'
      security:
      - bville_auth:
        - write_lama
        - read_lama
      x-codegen-request-body-name: body
components:
  schemas:
    Harmonogram:
      type: object
      properties:
        Warta:
          type: object
          properties:
            lamaID:
              type: integer
              format: int64
            poczatekWarty:
              type: string
              description: 'Date format according to ISO_8601 for example: yyyy-MM-dd''T''HH:mm:ss.SSSZ'
              format: date-time
            koniecWarty:
              type: string
              description: 'Date format according to ISO_8601 for example: yyyy-MM-dd''T''HH:mm:ss.SSSZ'
              format: date-time
    Lama:
      type: object
      properties:
        id:
          type: integer
          format: int64
        imie:
          type: string
        wiek:
          type: integer
          format: int64
        rasa:
          type: object
          properties: {}
        kolor:
          type: object
          properties: {}
        kubraczek:
          type: string
      example:
        id: 1234
        name: Sniezka
        wiek: 12
        rasa: lama
        kolor: kremowa
        kubraczek: podrozny
    CreateRequest:
      required:
      - lama
      type: object
      properties:
        lama:
          $ref: '#/components/schemas/Lama'
    UpdateRequest:
      required:
      - lama
      type: object
      properties:
        lama:
          $ref: '#/components/schemas/Lama'
    LlamaListResponse:
      required:
      - llamaList
      type: object
      properties:
        llamaList:
          type: array
          items:
            $ref: '#/components/schemas/Lama'
    AddScheduleRequest:
      required:
      - schedule
      type: object
      properties:
        schedule:
          $ref: '#/components/schemas/Harmonogram'
    ScheduleListResponse:
      type: object
      properties:
        llamaList:
          type: array
          items:
            $ref: '#/components/schemas/Harmonogram'
    Exception:
      type: object
      properties:
        code:
          type: integer
        detail:
          type: string
      example:
        code: 401
        name: Unauthorized
  securitySchemes:
    bville_auth:
      type: oauth2
      flows:
        password:
          tokenUrl: https://brzydakville.space:65536/token
          scopes:
            write_lama: modify llama for given guard stance
            read_lama: read llama data